# Optimize recursion of System.Net.Http.HttpRuleParser.GetExpressionLength