@MarcoRossignoli 2019-03-11
Unrelated outerloop fail https://mc.dot.net/#/user/dotnet-bot/pr~2Fdotnet~2Fcorefx~2Frefs~2Fpull~2F35959~2Fmerge/test~2Ffunctional~2Fcli~2F~2Fouterloop~2F/20190311.1/workItem/System.Net.Http.Functional.Tests/analysis/xunit/System.Net.Http.Functional.Tests.SocketsHttpHandler_HttpClientHandler_Proxy_Test~2FAuthProxy__ValidCreds_ProxySendsRequestToServer(proxyAuthScheme:%20Digest,%20secureServer:%20True,%20proxyClosesConnectionAfterFirst407Response:%20True)