@MarcoRossignoli 2019-03-12
/azp run corefx-ci (macOS x64_Debug)