@safern 2019-03-12
> @safern I tried to re-run CI because one leg fail(corefx-ci (macOS x64_Debug))...but seems that failed leg doesn't run, what I'm missing?

Running an individual leg through comment is not yet supported. We've requested this feature. I added it under the known issues section in the docs so that I submit a PR to update it whenever it is enabled:

https://github.com/dotnet/corefx/blob/master/Documentation/project-docs/pullrequest-builds.md#known-issues -- number 7